Julia API

OnlinePCA.jl (Julia API)

Binarization

OnlinePCA.csv2bin — Method.

csv2bin(;csvfile::AbstractString="", binfile::AbstractString="")

Convert a CSV file to Julia Binary file.

csvfile and binfile are specified such as Data.csv and Data.dat, respectively.

source

Summarization

OnlinePCA.sumr — Method.

sumr(;binfile::AbstractString="", outdir::AbstractString=".", pseudocount::Number=1.0)

Extract the summary information of data matrix.

binfile is a Julia Binary file generated by csv2bin function.

outdir is specified the directory you want to save the result.

pseudocount is specified to avoid NaN by log10(0) and used when Feature_LogMeans.csv <log10(mean+pseudocount) value of each feature> is generated.

Output Files

Sample_NoCounts.csv : Sum of counts in each column.

Feature_Means.csv : Mean in each row.

Feature_LogMeans.csv : Log10(Mean+pseudocount) in each row.

Feature_Vars.csv : Sample variance in each row.

Feature_CV2s.csv : Coefficient of Variation in each row.

Feature_NoZeros.csv : Number of zero-elements in each row.

source

Filtering

OnlinePCA.filtering — Method.

filtering(;input::AbstractString="", featurelist::AbstractString="", samplelist::AbstractString="", thr1::Number=0, thr2::Number=0, direct1::AbstractString="+", direct2::AbstractString="+", outdir::AbstractString=".")

This function filters the genes by some standards such as mean or variance of the genes.

Input Arguments

input : A Julia Binary file generated by csv2bin function.

featurelist : A row-wise summary data such as. The CSV files are generated by csv2bin function.

thr : The threshold to reject low-signal feature.

outdir : The directory specified the directory you want to save the result.

Output Files

filtered.zst : Filtered binary file.

source

Identifying Highly Variable Genes

OnlinePCA.hvg — Method.

hvg(;binfile::AbstractString="", rowmeanlist::AbstractString="", rowvarlist::AbstractString="", rowcv2list::AbstractString="", outdir::AbstractString=".")

This function perform highly variable genes, which is an feature selection method in scRNA-seq studies.

Input Arguments

binfile is a Julia Binary file generated by csv2bin function.

rowmeanlist : The mean of each row of matrix. The CSV file is generated by sumr functions.

rowvarlist : The variance of each row of matrix. The CSV file is generated by sumr functions.

rowcv2list : The cv2 of each row of matrix. The CSV file is generated by sumr functions.

outdir : The directory specified the directory you want to save the result.

Output Files

HVG_useForFit.csv : Parameters to estimate the highly variable genes.

HVG_a0.csv : Parameters to estimate the highly variable genes.

HVG_a1.csv : Parameters to estimate the highly variable genes.

HVG_afits.csv : Parameters to estimate the highly variable genes.

HVG_varFitRatio.csv : Parameters to estimate the highly variable genes.

HVG_df.csv : Parameters to estimate the highly variable genes.

HVG_pval.csv : Parameters to estimate the highly variable genes.

Reference

Highly Variable Genes

source

Oja

OnlinePCA.oja — Method.

oja(;input::AbstractString="", outdir::Union{Nothing,AbstractString}=nothing, scale::AbstractString="ftt", pseudocount::Number=1.0, rowmeanlist::AbstractString="", rowvarlist::AbstractString="", colsumlist::AbstractString="", dim::Number=3, stepsize::Number=0.1, numepoch::Number=3, scheduling::AbstractString="robbins-monro", g::Number=0.9, epsilon::Number=1.0e-8, lower::Number=0, upper::Number=1.0f+38, expvar::Number=0.1f0, evalfreq::Number=5000, offsetFull::Number=1f-20, offsetStoch::Number=1f-6, logdir::Union{Nothing,AbstractString}=nothing, perm::Bool=false)

Online PCA solved by stochastic gradient descent method, also known as Oja's method.

Input Arguments

input : Julia Binary file generated by OnlinePCA.csv2bin function.

outdir : The directory specified the directory you want to save the result.

scale : {log,ftt,raw}-scaling of the value.

pseudocount : The number specified to avoid NaN by log10(0) and used when Feature_LogMeans.csv <log10(mean+pseudocount) value of each feature> is generated.

rowmeanlist : The mean of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

rowvarlist : The variance of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

colsumlist : The sum of counts of each columns of matrix. The CSV file is generated by OnlinePCA.sumr functions.

dim : The number of dimension of PCA.

stepsize : The parameter used in every iteration.

numepoch : The number of epoch.

scheduling : Learning parameter scheduling. robbins-monro, momentum, nag, and adagrad are available.

g : The parameter that is used when scheduling is specified as nag.

epsilon : The parameter that is used when scheduling is specified as adagrad.

lower : Stopping Criteria (When the relative change of error is below this value, the calculation is terminated)

upper : Stopping Criteria (When the relative change of error is above this value, the calculation is terminated)

evalfreq : Evaluation Frequency of Reconstruction Error

offsetFull : Off set value for avoding overflow when calculating full gradient

offsetStoch : Off set value for avoding overflow when calculating stochastic gradient

logdir : The directory where intermediate files are saved, in every evalfreq (e.g. 5000) iteration.

perm : Whether the data matrix is shuffled at random.

Output Arguments

W : Eigen vectors of covariance matrix (No. columns of the data matrix × dim)

λ : Eigen values (dim × dim)

V : Loading vectors of covariance matrix (No. rows of the data matrix × dim)

Reference

SGD-PCA（Oja's method) : Erkki Oja et. al., 1985, Erkki Oja, 1992

source

CCIPCA

OnlinePCA.ccipca — Method.

ccipca(;input::AbstractString="", outdir::Union{Nothing,AbstractString}=nothing, scale::AbstractString="ftt", pseudocount::Number=1.0, rowmeanlist::AbstractString="", rowvarlist::AbstractString="",colsumlist::AbstractString="", dim::Number=3, stepsize::Number=0.1, numepoch::Number=3, lower::Number=0, upper::Number=1.0f+38, expvar::Number=0.1f0, evalfreq::Number=5000, offsetFull::Number=1f-20, offsetStoch::Number=1f-15, logdir::Union{Nothing,AbstractString}=nothing, perm::Bool=false)

Online PCA solved by candid covariance-free incremental PCA.

Input Arguments

input : Julia Binary file generated by OnlinePCA.csv2bin function.

outdir : The directory specified the directory you want to save the result.

scale : {log,ftt,raw}-scaling of the value.

pseudocount : The number specified to avoid NaN by log10(0) and used when Feature_LogMeans.csv <log10(mean+pseudocount) value of each feature> is generated.

rowmeanlist : The mean of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

rowvarlist : The variance of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

colsumlist : The sum of counts of each columns of matrix. The CSV file is generated by OnlinePCA.sumr functions.

dim : The number of dimension of PCA.

stepsize : The parameter used in every iteration.

numepoch : The number of epoch.

lower : Stopping Criteria (When the relative change of error is below this value, the calculation is terminated)

upper : Stopping Criteria (When the relative change of error is above this value, the calculation is terminated)

evalfreq : Evaluation Frequency of Reconstruction Error

offsetFull : Off set value for avoding overflow when calculating full gradient

offsetStoch : Off set value for avoding overflow when calculating stochastic gradient

logdir : The directory where intermediate files are saved, in every evalfreq (e.g. 5000) iteration.

perm : Whether the data matrix is shuffled at random.

Output Arguments

W : Eigen vectors of covariance matrix (No. columns of the data matrix × dim)

λ : Eigen values (dim × dim)

V : Loading vectors of covariance matrix (No. rows of the data matrix × dim)

Reference

CCIPCA : Juyang Weng et. al., 2003

source

GD

OnlinePCA.gd — Method.

gd(;input::AbstractString="", outdir::Union{Nothing,AbstractString}=nothing, scale::AbstractString="ftt", pseudocount::Number=1.0, rowmeanlist::AbstractString="", rowvarlist::AbstractString="",colsumlist::AbstractString="", dim::Number=3, stepsize::Number=0.1, numepoch::Number=3, scheduling::AbstractString="robbins-monro", g::Number=0.9, epsilon::Number=1.0e-8, lower::Number=0, upper::Number=1.0f+38, expvar::Number=0.1f0, evalfreq::Number=5000, offsetFull::Number=1f-20, offsetStoch::Number=1f-6, logdir::Union{Nothing,AbstractString}=nothing, perm::Bool=false)

Online PCA solved by gradient descent method.

The convergence is relatively slower than the other online PCA methods.

Input Arguments

input : Julia Binary file generated by OnlinePCA.csv2bin function.

outdir : The directory specified the directory you want to save the result.

scale : {log,ftt,raw}-scaling of the value.

pseudocount : The number specified to avoid NaN by log10(0) and used when Feature_LogMeans.csv <log10(mean+pseudocount) value of each feature> is generated.

rowmeanlist : The mean of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

rowvarlist : The variance of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

colsumlist : The sum of counts of each columns of matrix. The CSV file is generated by OnlinePCA.sumr functions.

dim : The number of dimension of PCA.

stepsize : The parameter used in every iteration.

numepoch : The number of epoch.

scheduling : Learning parameter scheduling. robbins-monro, momentum, nag, and adagrad are available.

g : The parameter that is used when scheduling is specified as nag.

epsilon : The parameter that is used when scheduling is specified as adagrad.

lower : Stopping Criteria (When the relative change of error is below this value, the calculation is terminated)

upper : Stopping Criteria (When the relative change of error is above this value, the calculation is terminated)

evalfreq : Evaluation Frequency of Reconstruction Error

offsetFull : Off set value for avoding overflow when calculating full gradient

offsetStoch : Off set value for avoding overflow when calculating stochastic gradient

logdir : The directory where intermediate files are saved, in every 1000 iteration.

perm : Whether the data matrix is shuffled at random.

Output Arguments

W : Eigen vectors of covariance matrix (No. columns of the data matrix × dim)

λ : Eigen values (dim × dim)

V : Loading vectors of covariance matrix (No. rows of the data matrix × dim)

source

RSGD

OnlinePCA.rsgd — Method.

rsgd(;input::AbstractString="", outdir::Union{Nothing,AbstractString}=nothing, scale::AbstractString="ftt", pseudocount::Number=1.0, rowmeanlist::AbstractString="", rowvarlist::AbstractString="",colsumlist::AbstractString="", dim::Number=3, stepsize::Number=0.1, numepoch::Number=3, scheduling::AbstractString="robbins-monro", g::Number=0.9, epsilon::Number=1.0e-8, lower::Number=0, upper::Number=1.0f+38, expvar::Number=0.1f0, evalfreq::Number=5000, offsetFull::Number=1f-20, offsetStoch::Number=1f-6, logdir::Union{Nothing,AbstractString}=nothing, perm::Bool=false)

Online PCA solved by Riemannian stochastic gradient descent method.

Input Arguments

input : Julia Binary file generated by OnlinePCA.csv2bin function.

outdir : The directory specified the directory you want to save the result.

scale : {log,ftt,raw}-scaling of the value.

pseudocount : The number specified to avoid NaN by log10(0) and used when Feature_LogMeans.csv <log10(mean+pseudocount) value of each feature> is generated.

rowmeanlist : The mean of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

rowvarlist : The variance of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

colsumlist : The sum of counts of each columns of matrix. The CSV file is generated by OnlinePCA.sumr functions.

dim : The number of dimension of PCA.

stepsize : The parameter used in every iteration.

numepoch : The number of epoch.

scheduling : Learning parameter scheduling. robbins-monro, momentum, nag, and adagrad are available.

g : The parameter that is used when scheduling is specified as nag.

epsilon : The parameter that is used when scheduling is specified as adagrad.

lower : Stopping Criteria (When the relative change of error is below this value, the calculation is terminated)

upper : Stopping Criteria (When the relative change of error is above this value, the calculation is terminated)

evalfreq : Evaluation Frequency of Reconstruction Error

offsetFull : Off set value for avoding overflow when calculating full gradient

offsetStoch : Off set value for avoding overflow when calculating stochastic gradient

logdir : The directory where intermediate files are saved, in every evalfreq (e.g. 5000) iteration.

perm : Whether the data matrix is shuffled at random.

Output Arguments

W : Eigen vectors of covariance matrix (No. columns of the data matrix × dim)

λ : Eigen values (dim × dim)

V : Loading vectors of covariance matrix (No. rows of the data matrix × dim)

Reference

RSGD-PCA : Silvere Bonnabel, 2013

source

SVRG

OnlinePCA.svrg — Method.

svrg(;input::AbstractString="", outdir::Union{Nothing,AbstractString}=nothing, scale::AbstractString="ftt", pseudocount::Number=1.0, rowmeanlist::AbstractString="", rowvarlist::AbstractString="", colsumlist::AbstractString="", dim::Number=3, stepsize::Number=0.1, numepoch::Number=3, scheduling::AbstractString="robbins-monro", g::Number=0.9, epsilon::Number=1.0e-8, lower::Number=0, upper::Number=1.0f+38, expvar::Number=0.1f0, evalfreq::Number=5000, offsetFull::Number=1f-20, offsetStoch::Number=1f-6, logdir::Union{Nothing,AbstractString}=nothing, perm::Bool=false)

Online PCA solved by variance-reduced stochastic gradient descent method, also known as VR-PCA.

Input Arguments

input : Julia Binary file generated by OnlinePCA.csv2bin function.

outdir : The directory specified the directory you want to save the result.

scale : {log,ftt,raw}-scaling of the value.

pseudocount : The number specified to avoid NaN by log10(0) and used when Feature_LogMeans.csv <log10(mean+pseudocount) value of each feature> is generated.

rowmeanlist : The mean of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

rowvarlist : The variance of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

colsumlist : The sum of counts of each columns of matrix. The CSV file is generated by OnlinePCA.sumr functions.

dim : The number of dimension of PCA.

stepsize : The parameter used in every iteration.

numepoch : The number of epoch.

scheduling : Learning parameter scheduling. robbins-monro, momentum, nag, and adagrad are available.

g : The parameter that is used when scheduling is specified as nag.

epsilon : The parameter that is used when scheduling is specified as adagrad.

lower : Stopping Criteria (When the relative change of error is below this value, the calculation is terminated)

upper : Stopping Criteria (When the relative change of error is above this value, the calculation is terminated)

evalfreq : Evaluation Frequency of Reconstruction Error

offsetFull : Off set value for avoding overflow when calculating full gradient

offsetStoch : Off set value for avoding overflow when calculating stochastic gradient

logdir : The directory where intermediate files are saved, in every evalfreq (e.g. 5000) iteration.

perm : Whether the data matrix is shuffled at random.

Output Arguments

W : Eigen vectors of covariance matrix (No. columns of the data matrix × dim)

λ : Eigen values (dim × dim)

V : Loading vectors of covariance matrix (No. rows of the data matrix × dim)

Reference

SVRG-PCA : Ohad Shamir, 2015

source

RSVRG

OnlinePCA.rsvrg — Method.

rsvrg(;input::AbstractString="", outdir::Union{Nothing,AbstractString}=nothing, scale::AbstractString="ftt", pseudocount::Number=1.0, rowmeanlist::AbstractString="", rowvarlist::AbstractString="",colsumlist::AbstractString="", dim::Number=3, stepsize::Number=0.1, numepoch::Number=3, scheduling::AbstractString="robbins-monro", g::Number=0.9, epsilon::Number=1.0e-8, lower::Number=0, upper::Number=1.0f+38, expvar::Number=0.1f0, evalfreq::Number=5000, offsetFull::Number=1f-20, offsetStoch::Number=1f-6, logdir::Union{Nothing,AbstractString}=nothing, perm::Bool=false)

Online PCA solved by Riemannian variance-reduced stochastic gradient descent method.

Input Arguments

input : Julia Binary file generated by OnlinePCA.csv2bin function.

outdir : The directory specified the directory you want to save the result.

scale : {log,ftt,raw}-scaling of the value.

pseudocount : The number specified to avoid NaN by log10(0) and used when Feature_LogMeans.csv <log10(mean+pseudocount) value of each feature> is generated.

rowmeanlist : The mean of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

rowvarlist : The variance of each row of matrix. The CSV file is generated by OnlinePCA.sumr functions.

colsumlist : The sum of counts of each columns of matrix. The CSV file is generated by OnlinePCA.sumr functions.

dim : The number of dimension of PCA.

stepsize : The parameter used in every iteration.

numepoch : The number of epoch.

scheduling : Learning parameter scheduling. robbins-monro, momentum, nag, and adagrad are available.

g : The parameter that is used when scheduling is specified as nag.

epsilon : The parameter that is used when scheduling is specified as adagrad.

lower : Stopping Criteria (When the relative change of error is below this value, the calculation is terminated)

upper : Stopping Criteria (When the relative change of error is above this value, the calculation is terminated)

evalfreq : Evaluation Frequency of Reconstruction Error

offsetFull : Off set value for avoding overflow when calculating full gradient

offsetStoch : Off set value for avoding overflow when calculating stochastic gradient

logdir : The directory where intermediate files are saved, in every evalfreq (e.g. 5000) iteration.

perm : Whether the data matrix is shuffled at random.

Output Arguments

W : Eigen vectors of covariance matrix (No. columns of the data matrix × dim)

λ : Eigen values (dim × dim)

V : Loading vectors of covariance matrix (No. rows of the data matrix × dim)

Reference

RSVRG-PCA : Hongyi Zhang, et. al., 2016, Hiroyuki Sato, et. al., 2017

source